+++ SalesWebMvc/SalesWebMvc/Migrations/SalesWebMvcContextModelSnapshot.cs
@@ -38,7 +38,7 @@ namespace SalesWebMvc.Migrations
                         .ValueGeneratedOnAdd()
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<decimal>("Amount");
+                    b.Property<double>("Amount");
 
                     b.Property<DateTime>("Date");
 
@@ -59,7 +59,7 @@ namespace SalesWebMvc.Migrations
                         .ValueGeneratedOnAdd()
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<decimal>("BaseSalary");
+                    b.Property<double>("BaseSalary");
 
                     b.Property<DateTime>("Birthday");
